refactor(mocks): improve predefined mocks API
Add the SccrollMockTrigger structure defining what mocks are expected to
do. Now the user give a pointer of this structure to sccroll_mockTrigger
which acts according to its values.

Add the SccrollMockOptions values to define some options for mocks. The
basic options are SCCMFLUSH which make the module forget the
current SccrollMockTrigger after the triggered mock call, and SCCMABORT
which aborts the program if a mock is called after one that (should
have) raised an error (allowing to better pinpoint unhandled cases).

Add the sccroll_mockFlush function to manually forget the current
SccrollMockTrigger.

Update unit tests.
Update documentation.

Signed-off-by: Alexandre Martos <contact@amartos.fr>