feat(assert)!: allow other signals for the sccroll_fatal functions
This commit changes the API of the sccroll_fatal family functions, as
the stream parameter has been replaced by sigint. However, this is more
logical, as these functions should only print on stderr (which is now
the default).

As any signal may be raised, the abort() redefinition is now partly
obsolete; this is made to ensure a __gcov_dump() call for SIGABRT.
A __gcov_dump() call must now be made in the fatal family functions,
thus the move of the prototype in the assert module header (which is
included by the mocks module).

BREAKING CHANGE: sccroll_fatal family API changed
Signed-off-by: Alexandre Martos <contact@amartos.fr>