feat(mocks)!: remove options for mocks
Options were complicating the whole mocks testing process.

Now the SCCMABORT option is applied by default, ie. the sccroll_mockFire
function aborts if any mock tested is not properly handled. As the
SCCMFLUSH option does not make sense on its own, it was also removed.

This removal lead to a huge simplification on the mocks triggering
process. Indeed, the whole point of the SccrollMockTrigger structure
was to handle the mock ID, the delay and the options. Now the function
sccroll_mockTrigger simply takes the mock and delay as parameters. This
also avoids the confusion between the structure and function with the
same name.

Internally, as the SccrollMockTrigger structure has disappeared, the
trigger variable (holding the informations) has been refactored to an
integer array, and the rest of the module adapted to it.

BREAKING CHANGE: removal of options and modifications of the mocks API
Signed-off-by: Alexandre Martos <contact@amartos.fr>