fix(mocks): rename functions for clarity
sccroll_mockAbortNotHandled was both a mouthful and a misleading
name. It is basically asserting that the triggered mock has been
handled, but sccroll_mockAssert was already taken. The latter is used
by sccroll_mockPredefined to test if a given wrapper handles properly
the errors triggered.

Thus, sccroll_mockAbortNotHandled has been renamed to
sccroll_mockAssert, and the other function renamed
sccroll_mockCrashTest.

Signed-off-by: Alexandre Martos <contact@amartos.fr>