fix(mocks)!: mocks affecting loaded shared libraries
The mocks were defined using the GCC linker wrap ability. However, this
is too efficient, as it affects **all** the fonctions, including those
in the shared libraries -- even the standard. Thus, trigerring errors
was becoming a huge pain, with headaches and all. The best example was
the raise of SIGSEGV signals when triggering the *alloc family at
various almost-random stages; these functions are heavily used, and thus
the previous mocks module backend was not able to control *where* the
trigger had to be done.

The mocks module has thus be refactored to fix this. Now the API has two
steps, the definition of a mock with a new SCCROLL_MOCK macro, and the
replacement of the mocked function with a defined macro in the tested
module's header. This ensures that only locally written code is
modified.

The exit function has also been rewritten to ensure the ability to check
during a successfull exit. This makes the API stricter as *any* call
non handled with an exit and failed status code is considered a test
failure and detected even if no other mocks is called after the
triggered one. This can be circumvented by using sccroll_fork and
checking the obtained signal value (SIGABRT being the signal used by
the library for errors).

BREAKING CHANGE: the whole mocks API has been rewritten
Signed-off-by: Alexandre Martos <contact@amartos.fr>