feat(mocks): add fopen and ferror mocks
The ferror mock is particular in the sens that as a f* function returns
an error, the stream is checked for error or EOF using ferror or feof.
Thus these must be coordinated with the other f* stream functions.

As feof does not raise any error, only the ferror function has been
mocked, and reacts for any other f* stream function too (therefore
the sccroll_mockHasFired macro).

Signed-off-by: Alexandre Martos <contact@amartos.fr>