feat(mocks): ignore ferror trigger if FILE funcs. do not trigger
Simulating errors of ferror does not make sense, hence the new ignored
state for SCCEFERROR.

This commit introduces the new sccroll_mockIsIgnored function, which
helps to indicate mocks for which triggers are ignored.

The sccroll_mockAssert and sccroll_mockFire functions had to be
adapted to accept mocks usually ignored, but still in sync with other
mocks. This is helped by the new sccroll_mockSyncedTrigger function,
which disguise the current mock ID code for the synced mock ID.

Signed-off-by: Alexandre Martos <contact@amartos.fr>